Added building name support
BuildingNames class should be able to resolve all building names exactly
as found in Daggerfall. There are a few assumptions that need to be more
thoroughly tested, but so far matching perfectly.
Added faction and name support to DaggerfallUnity singleton and content
reader.
FactionFile can now return FactionData from ID or map Name to ID.
Working on an experimental layout mapping process in
_ExperimentalCityLayout. This code will be deleted once research
completed.

+++ Assets/Scripts/API/DFBlock.cs
@@ -169,7 +169,8 @@ namespace DaggerfallConnect
             public DFLocation.BuildingData[] BuildingDataList;
 
             /// <summary>Unknown data.</summary>
-            internal Byte[] Section2UnknownData;
+            internal UInt32[] Section2UnknownData;
+            //internal Byte[] Section2UnknownData;
 
             /// <summary>Length of block record data in bytes (array is 32 records long, but only up to numBlockDataRecords have valid data).</summary>
             public Int32[] BlockDataSizes;
